+++ NEWS.md
@@ -12,6 +12,7 @@ This release simplifies the "key" structure. The nesting and crossing definition
 ## New features
 
 * Added new S3 generic `has_gaps()` to quickly check if there are implicit time gaps for each key in a tsibble.
+* New `new_data()` to produce the future of a tsibble.
 * An empty tsibble is now valid and will not raise an error, which is made easier to programme with.
 * New vignette on handling implicit missingness.